requirements.txt
tensorboard==2.9.0
scipy==1.8.0
tqdm==4.64.0
pyyaml==6.0.
matplotlib==3.5.2